Removes a bunch of functions that are not needed anymore.
